docs: add hazbot TBD and split Sidebar to its own subpath
- src/hazbot/TBD.md catalogs outstanding work (rule-sets blocked on
  missing defaults, stubbed impls, missing tabs, sheet quality, the
  missing student-facing UI, engine feature gaps, tooling) and lays
  out the plan for extracting the engine into its own npm package.
- src/hazbot/engine/README.md placeholder points at index.ts /
  types.ts / the spec as the API source-of-truth; full package
  README lands at extraction.
- Drop `Sidebar` / `SidebarProps` from the engine barrel; consumers
  now import from `../engine/sidebar` directly. Codifies the
  boundary so it doesn't drift before `package.json#exports` lands.
- Renumber ruleset 35 categories to fix the duplicate `id: 4`.
