+++ src/hazbot/engine/index.ts
@@ -12,5 +12,3 @@ export { ParseError } from "./parser";
 export { ENGINE_VERSION } from "./version";
 export { AnalysisEngineProvider, useAnalysisEngine } from "./react";
 export type { AnalysisEngineProviderProps, HookReturn } from "./react";
-export { Sidebar } from "./sidebar";
-export type { SidebarProps } from "./sidebar";
