+++ src/hazbot/rule-sets/35.ts
@@ -35,21 +35,21 @@ export const ruleSet35: RuleSet<WildfireDefaults> = {
       expression: "ranSimulation WITH ForestWAWOSuppression AND NOT UniformDroughtLevels",
     },
     {
-      id: 4,
+      id: 5,
       studentAction: "Ran the simulation with 2 vegetations but varied terrain between the zones",
       feedback: "Hazbot: This investigation compares forest types in the same terrain! Make sure both zones are mountains. Then run the model again.",
       visualFeedback: "Arrow pointing to Setup with the label: “Set terrain here.”",
       expression: "ranSimulation WITH ForestWAWOSuppression AND NOT UniformTerrainTypes",
     },
     {
-      id: 5,
+      id: 6,
       studentAction: "Ran the simulation with forest with and without suppression, but did not place two sparks (one in each zone)",
       feedback: "Hazbot: This model has two zones for comparison. Add one spark in each zone so you can compare wildfire spread.",
       visualFeedback: "Arrow pointing to spark tool with label: “Add a spark to each zone.” Use the box overlay.",
       expression: "ranSimulation WITH ForestWAWOSuppression AND UniformTerrainTypes AND UniformDroughtLevels AND NOT OneSparkPerZone",
     },
     {
-      id: 6,
+      id: 7,
       studentAction: "Ran the simulation, set zones with forest and forest w/ suppression,, and placed one spark in each zone",
       feedback: "Hazbot: Great job! You set up the model to compare two types of forest. You’re ready to answer the questions below.",
       visualFeedback: "Confetti animation or subtle celebratory visual",
